backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-dotenv>=1.0.1
google-generativeai>=0.6.0
sentence-transformers>=2.2.2
torch>=2.1.0
chromadb-1.2.1
pysqlite3-binary
